Fix list of add buttons for layoutGrid missing all the buttons after adding a block to one of them
The list was re-calculated every time an inner block was added using the `insertBlock` method. This means it tried to select them again.

However, once you add a inner block to the layout grid, the button ceases to exist and gives place to the actual block, and the list becomes outdated, and thus the sequence to access the right column number.

The solution is to calculate the number of buttons once when the columns are setup in the `setupColumns` method and cache it in a class instance variable.